pyproject.toml
[project]
name = "printers"
version = "0.1.0"
description = ""
authors = [{ name = "one-zero-eight", email = "https://t.me/one_zero_eight" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "authlib>=1.1.2,<2",
    "colorlog>=6.8.2,<7",
    "cryptography>=43.0.1,<44",
    "fastapi>=0.115.0,<0.116",
    "fastapi-swagger>=0.2.3,<0.3",
    "gunicorn>=23.0.0,<24",
    "httpx>=0.27.2,<0.28",
    "pre-commit>=3.8.0,<4",
    "pydantic>=2.9.2,<3",
    "ruff>=0.6.7,<0.7",
    "uvicorn>=0.31.0,<0.32",
    "pycups>=2.0.4,<3",
    "unoserver~=3.1",
    "python-multipart>=0.0.20,<0.0.21",
    "pyipp>=0.17.0",
    "aiogram>=3.20.0.post0",
    "pypdf2>=3.0.1",
]

[dependency-groups]
mongo = ["beanie>=1.26.0,<2", "motor>=3.6.0,<4"]

[tool.uv]
default-groups = ["mongo"]

[tool.ruff]
line-length = 120
lint.ignore = [
    # wildcard imports are fine
    "F403",
    "F405",
    "PLR",
]
lint.extend-select = ["I", "UP", "PL"]
target-version = "py312"
